@@ -8,9 +8,10 @@
     "start": "electron ."
   },
   "dependencies": {
+    "@astonish/ast-parser": "^0.0.0",
     "@astonish/parser-babel": "^0.0.0",
-    "astroturf": "^0.8.1",
-    "bootstrap": "^4.1.3",
+    "astroturf": "^0.8.4",
+    "bootstrap": "^4.2.1",
     "common-tags": "^1.8.0",
     "dexie": "^2.0.4",
     "dexie-relationships": "^1.2.10",
@@ -18,7 +19,7 @@
     "electron-debug": "^2.0.0",
     "electron-is-dev": "^1.0.1",
     "execa": "^1.0.0",
-    "immer": "^1.8.2",
+    "immer": "^1.11.0",
     "iter-tools": "^6.1.7",
     "monaco-editor": "^0.15.6",
     "pluralize": "^7.0.0",
@@ -32,35 +33,32 @@
     "signal-exit": "^3.0.2",
     "split-lines": "^2.0.0",
     "store": "^2.0.12",
-    "typesafe-actions": "^2.0.4",
+    "typesafe-actions": "^3.0.0",
     "uncontrollable": "^6.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.0",
-    "@babel/core": "^7.2.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
     "@rauschma/stringio": "^1.4.0",
     "@types/invariant": "^2.2.29",
-    "@types/lodash": "^4.14.119",
-    "@types/react": "^16.7.17",
+    "@types/lodash": "^4.14.120",
+    "@types/react": "^16.7.20",
     "@types/react-dom": "^16.0.11",
     "@types/reselect": "^2.2.0",
     "electron-devtools-installer": "^2.2.4",
-    "eslint": "^5.10.0",
-    "eslint-config-prettier": "^3.3.0",
+    "eslint": "^5.12.1",
+    "eslint-config-prettier": "^3.6.0",
     "fork-ts-checker-webpack-plugin": "^0.5.2",
-    "husky": "^1.2.1",
-    "lerna": "^3.6.0",
-    "lint-staged": "^8.1.0",
     "monaco-editor-webpack-plugin": "^1.7.0",
     "node-sass": "^4.11.0",
-    "prettier": "^1.15.3",
+    "prettier": "^1.16.0",
     "reselect": "^4.0.0",
     "resize-observer-polyfill": "^1.5.1",
-    "typescript": "^3.2.2",
-    "webpack": "^4.27.1",
+    "typescript": "^3.2.4",
+    "webpack": "^4.29.0",
     "webpack-atoms": "^9.1.0",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.10",
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.1.14",
     "webpack-node-externals": "^1.7.2"
   }
 }
