@@ -12,22 +12,31 @@
       "parsers/*"
     ]
   },
+  "workspace-sources": {
+    "@astonish/*": [
+      "parsers/*/src"
+    ]
+  },
   "devDependencies": {
-    "@babel/cli": "^7.1.2",
-    "@babel/core": "^7.1.2",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
     "@babel/preset-typescript": "^7.1.0",
+    "@typescript-eslint/eslint-plugin": "^1.0.0",
     "babel-eslint": "^10.0.1",
     "babel-preset-jason": "^6.0.1",
-    "eslint": "^5.8.0",
+    "eslint": "^5.12.1",
+    "eslint-config-4catalyzer-typescript": "^0.9.1",
     "eslint-config-jason": "^4.1.0",
-    "eslint-config-prettier": "^3.1.0",
+    "eslint-config-prettier": "^3.6.0",
+    "eslint-import-resolver-workspaces": "^1.0.3",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.11.1",
-    "husky": "^1.1.2",
-    "lerna": "^3.4.3",
-    "lint-staged": "^7.3.0",
-    "node-sass": "^4.9.4",
-    "prettier": "^1.14.3"
+    "eslint-plugin-react": "^7.12.4",
+    "husky": "^1.3.1",
+    "lerna": "^3.10.6",
+    "lint-staged": "^8.1.0",
+    "node-sass": "^4.11.0",
+    "prettier": "^1.16.0",
+    "typescript-workspace-plugin": "^2.0.1"
   }
 }
