@@ -1,8 +1,7 @@
 import { useCallback } from 'react'
 import Editor from './Editor'
-import { UPDATE_INPUT } from '../ActionTypes'
 import { useStoreDispatch, useStoreState } from '../store'
-import { updateInput } from '../actions'
+import { updateInput, updateAst } from '../actions'
 import * as React from 'react'
 const propTypes = {}
 
@@ -14,7 +13,8 @@ function InputPane() {
     async (value: string) => {
       const ast = await parser.parse(value)
 
-      dispatch(updateInput({ ast, value }))
+      dispatch(updateInput({ value }))
+      dispatch(updateAst({ ast }))
     },
     [parser]
   )
