@@ -1,6 +1,9 @@
 import createResource from './utils/createResource'
+import Parser from '@astonish/ast-parser/src'
+import { string } from 'prop-types'
 
-async function getParser(parser) {
+async function getParser(parser: string): Promise<Parser> {
+  // @ts-ignore
   return global.require(parser)
 }
 
