@@ -1,6 +1,6 @@
-import { updateInput, updateParserOptions } from '../actions'
+import { updateInput, updateParserOptions, updateAst } from '../actions'
 import { getType as t, ActionType } from 'typesafe-actions'
-import { Parser } from '../types'
+import Parser from '@astonish/ast-parser/src'
 
 export interface ParsingState {
   ast: any
@@ -12,7 +12,7 @@ export interface ParsingState {
 }
 
 export type ParsingAction = ActionType<
-  typeof updateInput | typeof updateParserOptions
+  typeof updateInput | typeof updateAst | typeof updateParserOptions
 >
 
 const initial = {
@@ -23,11 +23,12 @@ const initial = {
 
 export default (state: ParsingState = initial, action: ParsingAction) => {
   switch (action.type) {
+    case t(updateAst):
     case t(updateInput):
       return { ...state, ...action.payload }
     case t(updateParserOptions):
       return { ...state, options: action.payload.options }
   }
-  
+
   return state
 }
